move service structures
